requirements.txt
crewai
crewai_tools
crewai[tools]
langchain_google_genai
load_dotenv
langchain_groq
streamlit
pymupdf